Add API level security with user token to todo operations

+++ src/todo/todo.controller.js
@@ -1,26 +1,46 @@
 import Todo from './todo.model'
+import { AuthenticationError } from 'apollo-server';
 
 const todoController = {
-    getUsers: () => Todo.find({}),
-    getUser: async ({ id }) => {
+    getTodos: ({ user }) => {
+        if(!user)
+            throw new AuthenticationError('You must log in')
+
+        return Todo.find({})
+    },
+    getTodo: async ({ id }, { user }) => {
+        if(!user)
+            throw new AuthenticationError('You must log in')
+
         const todo = await Todo.findById(id);
         return todo
     },
-    addUser: ({ task, description }) => {        
+    addTodo: ({ task, description }, { user }) => {
+        if(!user)
+            throw new AuthenticationError('You must log in') 
+
         const todo = new Todo({ task, description })
         todo.save()
         return todo
     },
-    updateUser: async ({ id, task, description }) => {
-        const todo = await Todo.findById(id)        
+    updateTodo: async ({ id, task, description }, { user }) => {
+        if(!user)
+            throw new AuthenticationError('You must log in')
+
+        const todo = await Todo.findById(id)
+
         if(task) 
             todo.task = task
         if(description) 
             todo.description = description
+
         todo.save();
         return todo
     },
-    deleteUser: async ({ id }) => {
+    deleteTodo: async ({ id }, { user }) => {
+        if(!user)
+            throw new AuthenticationError('You must log in')
+
         const todo = await Todo.findByIdAndRemove(id)
         return todo
     }
